old reading in from file

+++ bitonicsort/bitonic.c
@@ -3,7 +3,7 @@
 
 #define MAX_SIZE 1000000
 
-void compAndSwap(int* a, int i, int j, int dir) {
+void compAndSwap(long long* a, int i, int j, int dir) {
     if ((a[i] > a[j] && dir == 1) || (a[i] < a[j] && dir == 0)) {
         int temp = a[i];
         a[i] = a[j];
@@ -11,7 +11,7 @@ void compAndSwap(int* a, int i, int j, int dir) {
     }
 }
 
-void bitonicMerge(int* a, int low, int count, int dir) {
+void bitonicMerge(long long* a, int low, int count, int dir) {
     if (count > 1) {
         int k = count / 2;
         for (int i = low; i < low + k; i++)
@@ -22,7 +22,7 @@ void bitonicMerge(int* a, int low, int count, int dir) {
     }
 }
 
-void bitonicSort(int* a, int low, int count, int dir) {
+void bitonicSort(long long* a, int low, int count, int dir) {
     if (count > 1) {
         int k = count / 2;
         
@@ -40,7 +40,7 @@ int main() {
         return 1;
     }
 
-    int a[MAX_SIZE];
+    long long a[MAX_SIZE];
     int count = 0;
     int num;
 
@@ -53,7 +53,7 @@ int main() {
 
     printf("Original array: ");
     for (int i = 0; i < count; i++) {
-        printf("%d ", a[i]);
+        printf("%lld ", a[i]);
     }
     printf("\n");
 
@@ -61,7 +61,7 @@ int main() {
 
     printf("Sorted array: ");
     for (int i = 0; i < count; i++) {
-        printf("%d ", a[i]);
+        printf("%lld ", a[i]);
     }
     printf("\n");
